Formally implemented workflow plan to ParallelJob.
git-svn-id: http://alma-dms/svn/sakura/trunk@680 9a79bd28-2a0b-46f8-beb7-14fc758f186d

+++ bench/e2eReduce/src/option_parser.h
@@ -17,7 +17,7 @@ struct CalibrationOptions {
 };
 
 struct FlaggingOptions {
-	int edge_channels;
+	size_t edge_channels;
 	float clipping_threshold;
 	bool do_clipping;
 };
